pyproject.toml
[tool.poetry]
name = "codemodpydanticv1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
libcst = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
black = "^23.10.0"
ruff = "^0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
